refactor(S-3.01): extract keychain glue to auth/keychain.rs (AC-002)
Move ENV_* constants, OAUTH_APP_HINT, resolve_credential,
resolve_oauth_app_credentials, and resolve_oauth_app_credentials_for_test
out of mod.rs into keychain.rs.

AC-002: keychain.rs contains only CLI-layer credential-resolution glue;
it delegates to crate::api::auth for all actual keyring operations and
does not call keyring::Entry directly (pre-split baseline was already
0 direct keyring::Entry calls in cli/auth.rs).

ENV_OAUTH_CLIENT_ID, ENV_OAUTH_CLIENT_SECRET, OAUTH_APP_HINT, and
resolve_oauth_app_credentials_for_test re-exported under #[cfg(test)]
for unit tests in mod.rs. 612 unit tests pass.
